requirements.txt
fastapi
uvicorn
pyjwt
httpx
torch
tenseal
joblib
numpy
requests
